Adds shared example testing for service name setting
