+++ spec/datadog/tracing/contrib/dalli/configuration/configuration_spec.rb
@@ -1,31 +1,6 @@
 require 'datadog/tracing/contrib/dalli/configuration/settings'
+require 'datadog/tracing/contrib/service_name_settings_examples'
 
 RSpec.describe Datadog::Tracing::Contrib::Dalli::Configuration::Settings do
-  describe 'Option `service_name`' do
-    context 'when with service_name' do # default to include base
-      it do
-        expect(described_class.new(service_name: 'test-service').service_name).to eq('test-service')
-      end
-    end
-
-    context 'when without service_name v0' do # default to include base
-      it do
-        with_modified_env DD_TRACE_SPAN_ATTRIBUTE_SCHEMA: 'v0' do
-          expect(described_class.new.service_name).to eq('memcached')
-        end
-      end
-    end
-
-    context 'when without service_name v1' do # default to include base
-      it do
-        with_modified_env DD_TRACE_SPAN_ATTRIBUTE_SCHEMA: 'v1' do
-          expect(described_class.new.service_name).to eq('rspec')
-        end
-      end
-    end
-  end
-
-  def with_modified_env(options = {}, &block)
-    ClimateControl.modify(options, &block)
-  end
+  it_behaves_like 'service name setting', 'memcached'
 end
